- Moved everything to one file - Reorganized everything into one structure.
